requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiohttp-socks==0.8.4